Created util package
