+++ core/src/main/scala/org/scalatra/util/StringUtil.scala
@@ -1,0 +1,14 @@
+package org.scalatra.util
+
+object StringUtil {
+
+  def splitFirst(content: String, delimiter: String): Option[(String, String)] = {
+    val i = content.indexOf(delimiter)
+    if(i > 0){
+      Some((content.substring(0, i), content.substring(i + delimiter.length, content.length)))
+    } else {
+      None
+    }
+  }
+
+}
